fix: display of detailed information has been corrected
